add renderable module to cell class
