+++ lib/cell.rb
@@ -1,7 +1,10 @@
+require "./lib/renderable"
+
 class Cell
+  include Renderable
+
   attr_reader :coordinate,
               :ship
-
   def initialize(coordinate)
     @coordinate = coordinate
     @ship = nil
@@ -25,31 +28,5 @@ class Cell
       @ship.hit 
     end
     @fired_upon = true
-  end
-
-  def render(param = false)
-    if param == false
-      if !fired_upon?
-        "."
-      elsif empty? && fired_upon?
-        "M"
-      elsif fired_upon? && !empty? && !@ship.sunk?
-        "H"
-      elsif fired_upon? && !empty? && @ship.sunk?
-        "X"
-      end
-    elsif param == true
-      if !fired_upon? && empty?
-        "."
-      elsif !fired_upon? && !empty?
-        "S"
-      elsif empty? && fired_upon?
-        "M"
-      elsif fired_upon? && !empty? && !@ship.sunk?
-        "H"
-      elsif fired_upon? && !empty? && @ship.sunk?
-        "X"
-      end
-    end
   end
 end
